Update library structure to enable inclusion of other CUDD modules.
